+++ ostap/parallel/tests/test_parallel_kisa.py
@@ -145,7 +145,7 @@ def MASS1  ( s ) : return s.mass
 # =============================================================================
 def test_kisa2 () :
        
-    from ostap.fitting.selectors import SelectorWithVars, Variable  
+    from ostap.fitting.pyselectors import SelectorWithVars, Variable  
     variables = [
         Variable   ( 'mass'  , 'mass(mu+mu-)' ,  3.09 , 3.11 ) , 
         Variable   ( 'c2dtf' , 'chi2(dtf)'    , -1    , 10   ) , 
@@ -203,7 +203,7 @@ def test_kisa2 () :
 ##     from  ostap.trees.funcs import H1DFunc
 ##     xh1 = H1DFunc ( histo = h1 , xvar = 'pt' ) 
     
-##     from ostap.fitting.selectors import SelectorWithVars, Variable  
+##     from ostap.fitting.pyselectors import SelectorWithVars, Variable  
 ##     variables = [
 ##         ##Variable ( 'mass1' , 'mass(mu+mu-)' , 2 , 4 , lambda s : s.mass ) , 
 ##         Variable   ( 'mass'  , 'mass(mu+mu-)' ,  3.09 , 3.11 ) ,