+++ ostap/tools/tests/test_tools_reweight2.py
@@ -226,8 +226,8 @@ else :
 # =============================================================================
 ## make reweighting iterations
 
-from   ostap.tools.reweight         import Weight, makeWeights,  WeightingPlot, W2Data  
-from   ostap.fitting.selectors      import SelectorWithVars, Variable 
+from   ostap.tools.reweight           import Weight, makeWeights,  WeightingPlot, W2Data  
+from   ostap.fitting.pyselectors      import SelectorWithVars, Variable 
 import ostap.parallel.parallel_fill
 
 # =============================================================================
@@ -393,5 +393,5 @@ mcy_density  .draw ('e1 same')
 time.sleep(10)
 
 # =============================================================================
-# The END 
+##                                                                      The END 
 # =============================================================================